+++ etc/t2.c
@@ -1,12 +1,12 @@
 #include <stdlib.h>
 
-void test(int *r, int *x, int n);
+void test(short *r, short *x, short n);
 
 int main(int argc, long *argv[])
 {
   int n = 3;
-  int *x = (int *)malloc(n * sizeof(int));
-  int *r = (int *)malloc(n * sizeof(int));
+  short *x = (short *)malloc(n * sizeof(short));
+  short *r = (short *)malloc(n * sizeof(short));
   test(r, x, n);
   free(x);
   free(r);
